adding file upload/download + PDF creation and deletion

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require('../models/product');
 const { validationResult } = require('express-validator/check')
+const fileHelper = require('../util/file');
 exports.getAddProduct = (req, res, next) => {
   
   res.render('admin/edit-product', {
@@ -15,22 +16,36 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+  if(!image){
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {title: title, price: price, description: description},
+      errorMessage: 'Attached File is not an image.',
+      validationErrors: []
+      
+    });
+  }
   const errors = validationResult(req);
   if(!errors.isEmpty()){
     return res.status(422).render('admin/edit-product', {
       pageTitle: 'Add Product',
-      path: '/admin/edit-product',
+      path: '/admin/add-product',
       editing: false,
       hasError: true,
-      product: {title: title, imageUrl: imageUrl, price: price, description: description},
+      product: {title: title, price: price, description: description},
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array()
       
     });
   }
+
+  const imageUrl = image.path;
   const product = new Product({
     title: title,
     price: price,
@@ -85,7 +100,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
   const errors = validationResult(req);
   if(!errors.isEmpty()){
@@ -94,7 +109,7 @@ exports.postEditProduct = (req, res, next) => {
       path: '/admin/edit-product',
       editing: false,
       hasError: true,
-      product: {title: updatedTitle, imageUrl: updatedImageUrl, price: updatedPrice, description: updatedDesc, _id: prodId},
+      product: {title: updatedTitle, price: updatedPrice, description: updatedDesc, _id: prodId},
       errorMessage: errors.array()[0].msg,
       validationErrors: errors.array()
       
@@ -109,7 +124,11 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
+      
       return product.save().then(result => {
         console.log('UPDATED PRODUCT!');
         res.redirect('/admin/products');
@@ -144,7 +163,13 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id: prodId, userId: req.user._id})
+  Product.findById(prodId).then(product => {
+    if(!product){
+      return next(new Error('Product not found'));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne({_id: prodId, userId: req.user._id})
+  })
     .then(() => {
       console.log('DESTROYED PRODUCT');
       res.redirect('/admin/products');
